refactor: Merge 'hook' to 'hooks'
